Add coalesce() function
Small helper function which is useful either to take the first non-missing
value, or to replace missing values with a given value (when passing a value known
not to be missing). Also usable in its broadcasted form, where a non-null
scalar argument will be automatically used in place of all nulls.
Inspired by the equivalent SQL and dplyr function.

Currently, broadcast returns a Vector{Any} rather than the more efficient
Vector{Union{Null, T}} when the result contains nulls. This will have to
be improved in Base.

The function is not exported (like replace() and skip()), but it could be
at some point since its name is unlikely to conflict with other packages.