Deprecate Missings.skip in favor of skipmissing
With the inclusion in Base, there will not necessarily be a Missings module
(and there are not many fans of this name anyway), so better use a name
which can be exported. Missings.fail and Missings.replace are kept as-is
since they are not exported, but should eventually be deprecated in favor
of generic iterators under Base.Iterators (possibly called fail and replace?).